# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gerenciamento de e-commerce",
        "title": "API de E-commerce",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Suporte API",
            "url": "http://www.suporte.com",
            "email": "suporte@api.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/produtos": {
            "get": {
                "description": "Retorna a lista completa de produtos cadastrados",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "produtos"
                ],
                "summary": "Lista todos os produtos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Produto"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Cria um novo produto no sistema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "produtos"
                ],
                "summary": "Adiciona um novo produto",
                "parameters": [
                    {
                        "description": "Dados do Produto",
                        "name": "produto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Produto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Dados inválidos",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Produto já existe",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/produtos/search": {
            "get": {
                "description": "Retorna os produtos cujos nomes correspondem ao parâmetro de busca",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "produtos"
                ],
                "summary": "Busca produtos por nome",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome ou parte do nome para busca",
                        "name": "nome",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Produto"
                            }
                        }
                    },
                    "400": {
                        "description": "Nome não pode ser vazio",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/produtos/{id}": {
            "get": {
                "description": "Retorna os detalhes de um produto específico",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "produtos"
                ],
                "summary": "Busca um produto por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Produto",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Produto"
                        }
                    },
                    "404": {
                        "description": "Produto não encontrado",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Atualiza os dados de um produto existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "produtos"
                ],
                "summary": "Atualiza um produto",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Produto",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados atualizados do Produto",
                        "name": "produto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Produto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Dados inválidos",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Produto não encontrado",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove um produto do sistema",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "produtos"
                ],
                "summary": "Remove um produto",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Produto",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Produto está em pedidos",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Produto não encontrado",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/produtos/{id}/estoque": {
            "patch": {
                "description": "Ajusta a quantidade em estoque de um produto (positivo para incrementar, negativo para decrementar)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "produtos"
                ],
                "summary": "Atualiza estoque",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Produto",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Quantidade para ajuste",
                        "name": "quantidade",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Quantidade inválida",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Produto não encontrado",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "model.Produto": {
            "type": "object",
            "properties": {
                "categoria": {
                    "type": "string"
                },
                "descricao": {
                    "type": "string"
                },
                "estoque": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "preco": {
                    "type": "number"
                }
            }
        }
    }
}